Login-Register Design
The completed front end for the login and register pages along with routing to the different pages
