+++ package.json
@@ -16,6 +16,7 @@
     "aws-amplify": "^6.11.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^7.1.1",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.0"
   },
