+++ src/components/SignIn/Signin.css
@@ -1,0 +1,122 @@
+/* General container styling */
+.signin-form-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  min-height: 100vh;
+  background-color: #f9f9f9;
+  font-family: Arial, sans-serif;
+}
+
+/* Header styling */
+.header-signin-form {
+  text-align: center;
+  margin-bottom: 1rem;
+}
+
+.header-signin-form h1 {
+  font-size: 2.5rem;
+  color: #333;
+  margin: 0;
+}
+
+/* Input fields container */
+.signin-input-fields {
+  width: 100%;
+  max-width: 400px;
+  padding: 1.5rem;
+  background: #fff;
+  border: 1px solid #ddd;
+  border-radius: 10px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+  box-sizing: border-box;
+}
+
+/* Label and input styling */
+.signin-form-input {
+  margin-bottom: 1.5rem; /* Spacing between input groups */
+}
+
+label {
+  font-weight: bold;
+  font-size: 1rem;
+  color: #555;
+  display: block;
+  margin-bottom: 0.5rem;
+}
+
+.input-signin {
+  width: 100%; /* Ensures the input spans the full width of the container */
+  padding: 10px;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  font-size: 1rem;
+  box-sizing: border-box; /* Includes padding in width calculation */
+}
+
+.input-signin:focus {
+  border-color: #007bff;
+  outline: none;
+  box-shadow: 0 0 4px rgba(0, 123, 255, 0.5);
+}
+
+/* Button styling */
+button {
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+.signin-button {
+  width: 100%;
+  padding: 10px;
+  background-color: #007bff;
+  color: #fff;
+  font-size: 1rem;
+  border: none;
+  border-radius: 5px;
+  margin-bottom: 1rem;
+}
+
+.signin-button:hover {
+  background-color: #0056b3;
+}
+
+/* Link styling */
+.toRegister-form-text {
+  display: block; 
+  text-align: center; 
+  color: #666; 
+  font-size: 0.9rem; 
+  text-decoration: none; 
+  margin-top: 1rem; 
+}
+
+.toRegister-form-text:hover {
+  color: #007bff; 
+  text-decoration: none; 
+}
+
+
+/* Responsive Design */
+@media (max-width: 600px) {
+  .signin-input-fields {
+    padding: 1rem;
+  }
+
+  .header-signin-form h1 {
+    font-size: 2rem;
+  }
+
+  label {
+    font-size: 0.9rem;
+  }
+
+  .input-signin {
+    font-size: 0.9rem;
+  }
+
+  button {
+    font-size: 0.9rem;
+  }
+}
